pyproject.toml
[tool.poetry]
name = "fastvapi"
version = "0.1.0"
description = "Wrap around OpenAI Realtime API"
authors = ["Alejandro Garcia <garciaae@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "2.32.3"
openai = "^1.51.0"
langchain = "^0.3.2"
redis = "^5.1.1"
uvicorn = "0.30.6"
fastapi = "0.115.0"
langchain-openai = "^0.2.2"
phonenumbers = "^8.13.47"
orjson = "^3.10.7"
aiohappyeyeballs = "2.4.0"
aiohttp = "3.10.6"
aiohttp-retry = "2.8.3"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.6.0"
async-timeout = "4.0.3"
attrs = "24.2.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
click = "8.1.7"
exceptiongroup = "1.2.2"
frozenlist = "1.4.1"
h11 = "0.14.0"
idna = "3.10"
multidict = "6.1.0"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pyjwt = "2.9.0"
python-dotenv = "1.0.1"
sniffio = "1.3.1"
starlette = "0.38.6"
twilio = "9.3.2"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
websockets = "13.1"
yarl = "1.12.1"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
ipdb = "0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
